Move parser test utilities and model fixtures into the parsers conftest.
